Adding GDAL to CI
This isn’t an ideal situation but the versions of GDAL available by default in Debian distributions don’t align with CentOS used by IT so it’s easier to find an image with the CentOS version and upgrade Pytohn within that.

In time we may want to build a custom image for use in CI to prevent the number of repeat installs etc.