Adding timestamp to DB dump files
To workaround the fact that the rolling file backups only track files by SHA1 checksum. Where backups don’t change because no data has changed this leads to duplicates.

Including the timestamp in the file resolves this but isn’t ideal. Long term we’ll track the SHA1sum and the created timestamp.