Refactoring public schema to magic_managed
In preparation for supporting non-MAGIC managed datasets (initially by ignoring them).

This is an initial step to separate out the resources that we manage, and backup, from those we don’t - either because they are ‘system’ or generic (such as Postgis views, QGIS layer styles) - or because they are managed by other group.

The initial use-case for this is layers managed by Field Ops.

This change is largely straightforward but will require recreating the databases in each deployed environment.

As part of this change, local database instances now use the same schemas as deployed instances.