Preparing to add non-managed layers
Changing from initial magic_managed term to controlled, as set out in #199.

This includes a general refactoring of SQL scripts to indicates properly files that templating (for example to set the database name).

For local development a set of ‘resolved’ files from these templates are available providing conventional names are used (as documented).

The folder these files are in has changed to better reflect that they configure the DB as a whole, rather than being exclusive to datasets.

Instructions on how to add/update/remove controlled schemas have been moved to the usage section.